requirements.txt
PyYAML>=5.4.1
jsonschema>=4.0.0
requests>=2.26.0
pytest>=7.0.0
jinja2>=3.0.0  # 用于HTML报告生成
